requirements.txt
asgiref==3.8.1
boto3==1.26.32
botocore==1.29.32
bson==0.5.10
certifi==2022.12.7
charset-normalizer==2.1.1
distlib==0.3.6
Django==5.1.2
django-allauth==65.0.2
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-storages==1.13.1
django-taggit==6.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
filelock==3.8.0
idna==3.4
jmespath==1.0.1
numpy==2.1.0
openpyxl==3.1.5
pandas==2.2.2
Pillow==10.0.1
platformdirs==2.5.2
psycopg2==2.9.9
PyJWT==2.6.0
PyMySQL==1.1.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2024.1
urllib3==1.26.13
virtualenv==20.16.5
